+++ ricaun.Nuke.Example.Tests/UnitTest.cs
@@ -1,0 +1,20 @@
+using NUnit.Framework;
+
+namespace ricaun.Nuke.Example.Tests
+{
+    public class UnitTest
+    {
+        [Test]
+        public void TestPass()
+        {
+            Assert.Pass();
+        }
+
+        [Explicit]
+        [Test]
+        public void TestFail()
+        {
+            Assert.Fail();
+        }
+    }
+}